master

+++ org.adempiere.base/src/org/compiere/model/MJournal.java
@@ -582,7 +582,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	private String validatePeriod(Timestamp dateAcct) {
 		// Get Period
 		MDocType dt = MDocType.get(getCtx(), getC_DocType_ID());
-		MPeriod period = (MPeriod) getC_Period();
+		MPeriod period = MPeriod.get(getC_Period_ID());
 		SimpleDateFormat dateFormat = DisplayType.getDateFormat(DisplayType.Date);
 
 		if (! period.isInPeriod(dateAcct)) {
@@ -936,9 +936,15 @@ public class MJournal extends X_GL_Journal implements DocAction
 		m_processMsg = ModelValidationEngine.get().fireDocValidate(this,ModelValidator.TIMING_BEFORE_REACTIVATE);
 		if (m_processMsg != null)
 			return false;	
-		
+
 		// teo_sarca - FR [ 1776045 ] Add ReActivate action to GL Journal
 		MPeriod.testPeriodOpen(getCtx(), getDateAcct(), getC_DocType_ID(), getAD_Org_ID());
+
+		if (!DocumentEngine.canReactivateThisDocType(getC_DocType_ID())) {
+			m_processMsg = Msg.getMsg(getCtx(), "DocTypeCannotBeReactivated", new Object[] {MDocType.get(getC_DocType_ID()).getNameTrl()});
+			return false;
+		}
+
 		MFactAcct.deleteEx(MJournal.Table_ID, get_ID(), get_TrxName());
 		setPosted(false);
 		setProcessed(false);
